[fx][tests] Add fx tests to CQ
Create GN build support for 'fx self-test' bash tests and add them
to the //tools:tests hierarchy so that it can be executed in CQ.

Note: This CL reuses GN integration that was created for the GN SDK
and thus it refers to it directly. In the future, fx should either
have its own GN glue or the one shared with GN SDK should be
extracted to a neutral and reusable location.

Change-Id: I242251abd74950fd80379b4ca558e1169cb51a8f
Multiply: host_x64/fx_script_tests: 40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399181
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Renato Mangini Dias <mangini@google.com>