[libc] Separate some utility code for getifaddrs into a header
The implementation of getifaddrs is going to move to fdio, but much of
the address copying logic will remain the same. This CL places that
shared code into a header, and adds a BUILD.gn file to share it with
fdio.

Fixed: 54214
Change-Id: I92f55573ae3726474ba8e69e63f22e6bc2ab9ff0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400025
Commit-Queue: George Kulakowski <kulakowski@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>