[fshost] Serve /data stats in inspect
This is currently done in the archivist, but this is a much better place
to do so now that we can expose inspect from fshost.

This removes the need of global-data in the archivist which will be done
in a follow-up once clients of this data have been updated to use the
data exposed here instead.

Bug: 52951
Tested: fx shell runtests -n fshost-inspect-test-test

Change-Id: I1adc1bb042685df0b2492dcd3c476608990c457a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395360
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>