[IDK sysroot] Removing syscall FIDL files from headers list.
Currently the .fidl files in //zircon/vdso are being added
to the sysroot header file list with a relative path of ../../vdso/.
This breaks the Rust tar processing which does not allow ParentDir
relative paths in the tarball manifest.

This also creates an undocumented directory in the SDK at
sdk://arch/$cpu/vdso.

This change removes these files from the sysroot SDK atom.

Once the tooling is ready to be exposed for processing these
FIDL files, they should be added back as a fidl library atom and not
part of the sysroot SDK package.

The GN SDK generation script is also updated to accommodate the
new sysroot metadata.

Bug: 54116

Change-Id: I284240a949e9297d8bac4e8c2612dbbfa4243017
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397364
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>