[flatland] Add Flatland::DeregisterBufferCollection().
The client call to deregister a buffer collection only erases the
mapping from the client-generated ID to the Renderer-generated ID.
This prevents future API calls using the ID and allows the ID to be
reused, but does not change any Flatland-internal logic.

The actual garbage collection of the resources cannot occur until the
buffer collection is completely unused in the Flatland instance and
not inolved in any currrent or pending renders.

To determine the former, each buffer collection now has an associated
usage count, which is the number of Images that reference it.
Flatland increments this count when the client creates a new Image
using that buffer collection and decrements it when the Image is
garbage collected (not just when it is released).

To determine the latter, Flatland creates  a release fence to signal
when Renderer resources can be cleaned up.

In the future, the FrameScheduler will signal this fence. For now,
Flatland stores it and provides a function to signal all pending
Waits for test purposes. Issue 53330 tracks progress for removing
this.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: I8556814cb04f83fe19701f38d2021063214af810
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395904
Reviewed-by: Adam Gousetis <adamgousetis@google.com>
API-Review: Adam Gousetis <adamgousetis@google.com>
Testability-Review: Evan Birenbaum <birenbaum@google.com>
Commit-Queue: Evan Birenbaum <birenbaum@google.com>