[fidl] Enable subset of {encoding,decoding,validating} tests on host
When working with the walker it hard to debug because there is a cycle
where you can't boot but the tests depend on you booting to run them.

This CL enables a subset of tests on host to avoid this cycle as much as
possible.

Test: fx build, then out/default/host_x64-asan/exe.unstripped/fidl

Change-Id: If3db64e53123a4c41d7600d2c5f4ab950ed1b6e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400356
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: André Pang <apang@google.com>
Testability-Review: André Pang <apang@google.com>