[scheduler] Call MigrateFn after updating state
In a recent change, we begun using Thread::LastCpuLocked() in migration
functions after CPU migration. This is used to update the CPU that the
thread is currently executing from, so that we can target IPIs
correctly.

However, the scheduler was not updating the thread's SchedulerState
until after migration, therefore causing us to IPI the wrong CPU. This
can result in a dramatic slow down of the virtualization integration
tests, especially on arm64.

Change-Id: I4bf16cac89a9e30a06d177fe2d707e0335b960a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400253
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>