[kernel] Add internal ZX_JOB_NO_CHILDREN signal to jobs.
Currently Job objects expose the ZX_JOB_NO_JOBS and ZX_JOB_NO_PROCESSES
signal to indicate that the job has no child jobs or child processes,
respectively. However, there is currently no way for watchers to be only
notified when the job has no children (i.e., no jobs and no processes)
at all.

The current work-around is to listen for one signal and, if it fires,
check for the other. If the second signal is not active, we wait on it
instead. We may have to keep going back and forth as new jobs/processes
are created or destroyed:

  zx_signals_t remaining_signals = ZX_JOB_NO_JOBS | ZX_JOB_NO_PROCESSES;
  do {
    zx_signals_t observed;
    zx_object_wait_one(target_job, remaining_signals,
                       ZX_TIME_INFINITE, &observed);
    remaining_signals = (ZX_JOB_NO_JOBS | ZX_JOB_NO_PROCESSES) & ~observed;
  } while (remaining_signals != 0);

This CL adds a ZX_JOB_NO_CHILDREN signal, which is active when the job
has no children at all.

Currently, we only expose the constant in-kernel, allowing the
RootJobObserver to use it in an upcoming CL.

fxb/53986 tracks either making this signal visible to userspace, or
providing a more general mechanism to userspace avoiding the need for it
altogether.

Bug: 53986
Change-Id: I8436036a6cf5a358d4bbb34fe5df9495c009b8b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395756
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>