[components] ElfRunner monitors for process exit
The ElfRunner waits for a ZX_PROCESS_TERMINATED signal from the
process handle for the process the runner started to host the
component. When this signal is received, it aborts the future
serving the ComponentController channel to component_manager
which causes the channel to close.

Fixed: 49303

Change-Id: I44da97eac0c1b4966c4aa754d9e2db2c880087df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396072
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Gary Bressler <geb@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>