[audio][test] Fix WaitForPacket
After an AudioRenderer receives a Packet via SendPacket, the renderer
will hold onto the packet until it's done with that packet, at which
point the packet is returned to the caller.

It turns out that RendererShim::WaitForPacket was assuming that a packet
was completely rendered to the ring buffer by the time it is returned.
However, that may not be the case -- it may be mixed into an
intermediate buffer but not yet rendered to the final ring buffer. This
breaks tests (in a future CL) which rely on WaitForPacket to wait until
the packet is actually rendered to the ring buffer.

To fix this, I've refactored WaitForPackets such that it has enough
information to wait until the packet is fully rendered. Specifically, we
pass in the renderer's start_time (wall clock time for PTS=0) and each
packet's start_pts and size. From this info, we can compute the real
time after which the packet must have been rendered to the ring buffer.

Multiply: audio_pipeline_tests
Bugs: 49807
Fixed: 53893
Change-Id: I9f844b59e30cd07564750fe4a483027a67e7fb9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396999
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Martin Puryear <mpuryear@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>