[hypervisor][arm64] Fix use of thread pointer
Just as with the x86 code, we need to make sure we are careful when we
use a thread pointer, as it could result in a use-after-free.

Update the arm64 code to track the current CPU that the thread is
executing on, just as in the x86 code.

Also make some minor tweaks to align the arm64 and x86 code.

Change-Id: I674397bc163239aaee810d028346762f1bc085a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400254
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>