[fbl][slab_alloc] Fix a failure to passthru to global delete
Fix an issue with MANUAL_DELETE slab allocatable objects where the
hidden custom delete operator was failing to passthru to the
underlying global delete operator in the case that the
AllowManualDeleteOperator options was passed.

Typically, a slab allocated object whose slab allocator uses the
MANUAL_DELETE flavor should never have `delete` called on it.
Instead, the Delete method of it's slab allocator should be used
instead.  In order to prevent accidents, the delete operator of the
SlabAllocated<> base class is overloaded and placed in the protected
section of the base class.  Additionally, a runtime debug assert is
present to explode if somehow the operator gets invoked at runtime.

In a recent change

https://fuchsia-review.googlesource.com/c/fuchsia/+/390793

An option was added which (when passed to the traits of the object)
suppresses the assert and allows objects to have `delete` used on
them.  This allows for objects which can have a dual-life.  One where
they are sometimes slab allocated, but another where their lifecycle
is managed using the heap along with new/delete.  'delete' still needs
to be invoked either from a friend of the object or the object itself,
but it can be done and is considered to be an advanced use case.

That said, the SlabAllocated<> base class still needs to pass the
memory onto the global delete operator, otherwise the memory will be
leaked.

A test was written when the feature was added, but unfortunately it
has no good way to know that the global delete operator was actually
called.  For that, you need a tool like LSAN which is not currently
enabled in test environments, although it will be eventually, as soon
as we can make it to a green build/test point.

Enter bug #54182.  When LSAN was turned on, it was finding these
leaks, but in a different place.  Specifically, the RegionAllocator
when configure in no-slab-allocator mode (the first user of the
AllowManualDeleteOption).  In theory, it would also have found
the trouble when running the specific slab allocator test, it was just
that this bug was filed first.

There is a CL on deck which will be the first production user of the
feature

https://fuchsia-review.googlesource.com/c/fuchsia/+/391865

but it has not landed yet, and the allocations it performs live for
the entire life of the kernel, so thankfully there has be no
real-world damage yet.

Still, this is a serious bug and needs to be fixed, so the key
takeaway here is "Go LSAN!"

Tests: existing unit tests coupled with LSAN will prevent regressions
       here.
Fixed: 54182
Change-Id: I9d3562f8d2298159a38bf5e3c2f6dd5e31ecbfab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399627
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Christopher Anderson <cja@google.com>