[kernel][vm] Handle ZX_ERR_INTERNAL_INTR_KILLED in page fault
If the faulting thread is killed while waiting for a page request to be
fulfilled, return successfully from the page fault handler. There's
nothing we can do in this case. This is not indicative of an error on
the page fault path.

Test: core-pager

Bug: 54361
Change-Id: I285d96c3f531b6478b9a4f116403304b7ca014f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400582
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>