[debugger] Address a theoretical thread suspension race.
When a thread is requested to synchronously suspend, but there was
already an asynchronous suspend pending, wait until the suspend is
complete.

I noticed this theoretical race condition issue when looking at the
bookmark "stop mode" handling. Currently we never issue asynchronous
suspends so this issue never happens (there is one call to asynchronous
suspend, but that caller immediately waits for the thread to be
suspended). If we use suspend in more ways in the future, this might
become an issue so is worth resolving.

Change-Id: I508fc73603431a50512bb0fb5727b294d20209a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400723
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>