[xhci-rewrite] Add physically contiguous TRB allocator
This patch adds a new allocator which ensures that all TRBs
returned are contiguous in physical memory.
If we are in the middle of a transfer ring and there isn't
enough contiguous space, it pads the unusable space with NOPs.

Test: fx test xhci-transfer-ring-unittest
Change-Id: I1bbc2fb38a4f0b98105f5861bc1bf75c16917543
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400496
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>