[debugger] Add conditional expression support.
Adds support for C and Rust "if" statements, and C ternary ?:
expressions.

C-style if statements are not useful by themselves because our
expression system has no need to execute a sequence of steps (there is
no state). But if *expressions* (C ternary "?:" and Rust "if"
expressions) are useful for implementing both pretty-printers and, in
the future, conditional breakpoints and logging.

Change-Id: I46889503ff35af18d5b11d8f1a6c3f6c06300eca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398597
Testability-Review: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>