[vm][test] Fix flake in core-pager test
The CleanThreadKill test verifies that we can cleanly kill a thread
blocked on a page fault. One of the conditions it checks is that the
pager receives a page request. In some execution flows, the faulting
thread was getting killed before this check, which would also cancel the
page request, causing the test to hang waiting for a page request to
arrive.

Changed the control flow to make the sequence of events deterministic.
Got rid of a redundant thread.

Multiply: /boot/test/core-pager

Bug: 54468
Change-Id: I68144736b36b52fd73c2f4d3ebf182a8d727698c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401019
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>