[bt][avrcp] Fix peer disconnection logic
Currently, when a peer disconnects from AVRCP, the termination of the
AVCTP control channel task does not necessarily occur immediately. This
is because there can be outstanding requests (particularly the
hanging-get OnVolumeChanged request). Consequently, when the same peer
reconnects (i.e is not "forgotten" by the BT stack), a new AVCTP control
stream processing task is not spawned, because the original task is still
lingering.

It is expected behavior for the AVCTP control channel task to linger,
because of the outstanding hanging-get request. However, in the event
the same peer disconnects and reconnects, a new control channel
processing task should be spawned, and the old one should be cleared and
terminated.

Adds a new flag that when set, will terminate any outstanding tasks
before spawning the new task.

While we're here, update some of the log messages to be more useful.

Bug: 54168
Test: Added test_peer_reconnection

Multiply: bt-avrcp-tests

Change-Id: I9db223eff88eeb34c6c6ce5b9511169963736ecc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399134
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>