[logs] update diagnostic/logger FIDLs for user control
This change lays the foundation for the implementation of
dynamic runtime log level control, which will be implemented
as an event (specifically - a change in Interest, which
connotes a severity level) on the LogSink protocol.

The Controller protocol is removed as there are no immediate
plans for it.

Bug: 54198
Test: CI/CQ to ensure no regression
      FIDL only with no implementators yet, so no specific testing

Change-Id: Iea0629bfb863f1b5a7358b10e00458b076f13011
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399537
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Todd Lee <leetodd@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>