[msd-arm-mali] Prefer running atoms in current mode
Switching from protected to nonprotected mode takes around 500
microseconds. The code currently tries to schedule a mode switch as soon
as a it sees a runnable atom of the opposite type, so that can add up
quickly if there are very small command buffers.

Instead, prefer to execute atoms in the current mode (up to a limit of
20 in a row). This code won't reorder atoms within a job slot, but it
will select which job slot to use based on the current mode.  That's
generally enough, since if we're switching a lot then job slot 0 will be
in protected mode and job slot 1 (vertex shaders) will be in unprotected
mode.

Bug: b/159041726
Fixed: 13132

Change-Id: I59008c78b4ccbb92543ff313e39c2dd39301dd09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400717
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Testability-Review: Craig Stout <cstout@google.com>