[testrunner] add ability to test over serial
This option would be exercised if FUCHSIA_SSH_KEY is not found in the
environment and FUCHSIA_SERIAL_SOCKET is: a 'serial tester' would be
initialized, which would be a thin wrapper around a client socket
connection.

This change also removes the use of ioutil.Discard (which is /dev/null)
as it was discovered that writing to that was a source of significant
slowdown.

Bug: 41930

Change-Id: I69c032e851313478f5f7c4efba704a5f33e56810
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384014
Testability-Review: Gary Miguel <garymm@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>