[kernel][pcie] Avoid a segfault
Don't crash in the `k pciregions` kernel debug console command if the
PCIe driver has not been initialized.  Also, add a debug printf to
this command as well as the others to indicate the specific reason
that the command has done nothing in the case that the driver has not
been initialized when the command was run.

Change-Id: I250b91e049c18743d5055951c99229cbf2aa465e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400995
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>