[netstack] Stream socket shutdown fixes.
This change fixes 3 bugs in handling application stream socket shutdown:
(1) With the current implementation, on a shutdown, we update the
Zircon socket state and invoke gVisor endpoint shutdown lazily, as we
see read/write failure to Zircon socket. This breaks the SHUT_RDWR case,
where the TCP is expected to reply with a RST when it receives any data
from the remote post RDWR shutdown, because the gVisor endpoint would
not yet be notified of read shutdown from the application.

(2) The application poll currently does not see a POLLHUP when the
socket in CLOSE_WAIT receives a RST. The event is ignored because
loopRead would have returned on ClosedForReceive error.

(3) loop{Read,Write} goroutines can deadlock when both simultaneously
call endpointWithSocket.close(). This is because of the sync.Once.Do()
behavior where it waits for the first call to return before it lets
subsequent calls to bypass execution.

To address (1), we define a watcher goroutine per stream socket to poll
for zircon socket state changes to notify gVisor sooner and with all
state updates together. For (2), we ensure that we listen on error or
closing events from endpoint and trigger socket state close sooner.
For (3) we use compare-and-swap instead of sync.Once.

Bug: 38892
Test: fx test -od "fuchsia-pkg://fuchsia.com/netstack_c_api_tests#meta/netstack_bsdsocket_c_api_test.cmx" -- --gtest_filter=NetStreamSocketsTest.ShutdownReset --gtest_repeat=100000
      fx run-host-tests netstack_bsdsocket_c_api_test

Change-Id: Icbbf4c1941c754522211efb17e25a82ae60d8be6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400720
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>