[bt][sm] Clarify the state machine in the ScStage1 classes
Firstly, this CL cleans up a possible externally-triggered assertion in
the ScStage1JustWorksNumericComparison class from receiving multiple
Pairing Random values. The first Pairing Random message would, through
a series of callbacks, trigger the destruction of the ScStage1JustWorks
NumericComparison class by its owning class, so in practice, this would
not be exploitable. However, this depends on logic external to the
class, which is not desirable, so a check for this case has been added.
Secondly (and relatedly),this CL adds documentation and improves naming
to make the state flow of the ScStage1 classes more clear.

Bug: 53589
Test: Added SMP_ScStage1JustWorksNumericComparisonTest.ResponderReceive
      RandomTwiceFails to `$ fx test bt-host-unittests`
Change-Id: Ia517efffa18282ec9dd226ec940987a882cdae08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397367
Reviewed-by: Yu Shan <shanyu@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>