[kernel] Zero the symbolizer file VMO if not publishing it
All of the instrumentation data VMOs are always published to
userboot for simplicity, but zero-length VMOs will be ignored in
userland.  The symbolizer file VMO was left live with a size of
PAGE_SIZE when it wasn't meant to be published, so userland saw
it as live but its name hadn't been set.  Instead, just reset it
to zero size so it will be ignored like all the others.

Fixed: 54904
Change-Id: I9c7b49f070d3a6b4a02e065a99e50468aebcd503
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401607
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>