[wlan] Remove the definition of WlanInfoMacRoleMask
Adding WlanInfoMacRoleMask was a mistake. The enum type
generated by Banjo is intentionally overloaded for bitmasks in Rust
and C++. In general, adding *Mask types will not work in Rust code
because Rust sees the *Mask type as different from its sister * type.
The semantic benefit of adding the *Mask is very small and is
therefore removed.

Note that in the case of WlanInfoMacRoleMask, there are no
issues compiling. There are however issues with creating a type such
as WlanInfoHardwareCapabilityMask because Rust performs type checking
on case statements using the type and does not find any values for it.

Bug: 52941
Test: No additional tests necessary
Change-Id: Iaad67bee4a89ec883a8ed297db52e7cf9eda517c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400811
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Eric Wang <eyw@google.com>
Testability-Review: Eric Wang <eyw@google.com>