[fidl] Move non-nullable but absent logic to visitor
This removes kAllowNonNullableCollectionsToBeAbsent and replaces it with
a visitor method that is called when a non-nullable absent collection is
encountered.

This has a few effects:
- VisitPointer now only handles present pointers
- Separates out the special case of null string/vectors in LLCPP and
makes clearer what happens in this case.
- Opens the door for code size reductions in encoding.cc if ever needed,
because the linearize_and_encode and encode cases can be controlled by
run-time if statements. That said, this probably isn't advisable in
practice because it is likely that we get performance benefits by not
needing to check whether linearization should be performed at each step.

Bug: 53258
Test: Uses existing tests, e.g. fx test fidl-tests

Change-Id: Idc053096646a534ebd579878ef44898294d92cfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401913
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>