[tee_manager] Remove TeeDeviceConnection
The TeeDeviceConnection struct wrapped a DeviceConnectorProxy with an
AbortContext to enable registrants to abort their futures if the
channel to the DeviceConnectorProxy was closed. There was one
abort registrant for the ServiceFs and an additional abort registrant
for each passthrough connection to the tee.Device served by the device
driver. This CL restructures the tee_manager to not use
TeeDeviceConnection or AbortContexts to instead just do a select! on the
services future and the DeviceConnectorEventStream future so that the
tee_manager can shutdown appropriately.

Fixed: 51733

Change-Id: Iab61b67ecc98edcf83b550420aa16cdcbc697a3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399154
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>