[kernel][vm] Correct resizing of COW clones
This fixes a bug in the ReleaseCowParentPagesLocked where parent_limit_
could not be updated correctly. The failure to update parent_limit_
can be observed by
 * Shrinking a VMO in a hierarchy such that the bug is hit resulting in
   parent_limit_ being *larger* than the final VMO size
 * Resize the VMO larger. This causes page lookups that use
   parent_limit_ to start seeing parent data when they should not.

The previous ReleaseCowParentPagesLocked tried to be clever in
optimally walking up and down the parent hierarchy a minimal number of
times. Whilst this algorithm might have been fixable, or at the least
a correct algorithm of similar complexity found, I chose to use a
simpler approach. Although simpler, and of higher worst case complexity,
as these cases are unlikely to exist in practice having a more
obviously correct algorithm was more important. It can be re-optimized
in the future should it be a problem.

Test: The included test can fail in a few different ways without this
change
 1. If the extra assertion in Resize is included then the test will
    trip that assertion.
 2. The test would fail to read zeroes when it is suppose to after
    resizing larger.
 3. When writing it would trigger the assertion failure and backtrace
    in the referenced bug.
Although ReleaseCowParentPagesLocked is also called when
closing/merging children of hiddren parents I did not attempt to
find if there were any similar or other bugs possible through that
path. If there were bugs possible this should also fix them though.

Bug: 53710

Change-Id: I75af4c89df8595c21015bf7ccf16c35086a39078
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400262
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>