[blobfs] Wire up blobfs Inspect API
This CL wires up the blobfs Inspect API and
connects it to the diagnostics directory of fshost.

Archivist picks up blobfs and fshost diagnostics
and shows both with the `iquery show bootstrap/fshost`
command.

This is done by creating a RemoteDir subdirectory in the
fshost diagnostics directory and passing the server
end as an additional handle to the blobfs process.

blobfs::Runner serves the diagnostics directory
in addition to the root.

The additional handle passed in to blobfs is a
TEMPORARY measure. See fxb/54521 for more details

Test: 1. `iquery show bootstrap/fshost` now shows two
         trees, one for fshost and one for blobfs.
      2. Ran all tests -> `runtests /system/test/blobfs*`
Change-Id: Ibd2a17b1ca9a88bbe04f5d93860e92d49d1cd327
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400753
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>